+++ backend/src/main/java/com/example/demo/entity/UserEntity.java
@@ -3,13 +3,12 @@ package com.example.demo.entity;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import lombok.Setter;
 
 @Getter
 @AllArgsConstructor
 @RequiredArgsConstructor
 public class UserEntity {
-    private int id;
+    private long id;
     private String firstName;
     private String lastName;
     private String initial;
